Move reth_eth_wire::DisconnectReason to reth-eth-wire-types
